1.7.6

+++ controllers/notes/note.php
@@ -13,10 +13,15 @@ if($_SERVER['REQUEST_METHOD'] === 'POST') {
         'id' => $_GET['id']
     ])->findOrFail();
 
+    authorize($note['user_id'] === $currentUserId);
+
 
     $db->query('delete from notes where id = :id', [
         'id' => $_GET['id']
     ]);
+
+    header('location: /notes');
+    exit();
 } else {
 
     $note = $db->query('select * from notes where id = :id', [
